Move tests at root to the right places
